requirements.txt
python-fasthtml
requests
python-dotenv
